+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -13,6 +13,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 - Fixed the name of the Key Vault Managed HSM settings get command from `azmcp_keyvault_admin_get` to `azmcp_keyvault_admin_settings_get`. [[#643](https://github.com/microsoft/mcp/issues/643)]
 - Removed redundant DI instantiation of MCP server providers, as these are expected to be instantiated by the MCP server discovery mechanism. [[644](https://github.com/microsoft/mcp/pull/644)]
 - Fixed App Lens having a runtime error for reflection-based serialization when using native AoT MCP build. [[#639](https://github.com/microsoft/mcp/pull/639)]
+- Added validation for the PostgreSQL database query command `azmcp_postgres_database_query`.[[#518](https://github.com/microsoft/mcp/pull/518)]
 
 ### Other Changes
 - Updated the description of Service bus commands to decrease ambiguity and increase selection accuracy by LLMs: [[#642](https://github.com/microsoft/mcp/pull/642)]